Refactor the many proc closures.
