+++ src/lib.rs
@@ -5,11 +5,11 @@ use crate::os::OS;
 
 mod async_io;
 mod binfs;
-mod closures;
 mod compilation_mode;
 mod js;
 mod os;
 mod proc;
+mod proc_closures;
 mod term;
 
 #[wasm_bindgen]
